% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AfsariCorcpp}
\alias{AfsariCorcpp}
\title{Describe the function}
\usage{
AfsariCorcpp(x, mydist)
}
\arguments{
\item{describe}{the params}
}
\value{
describe the return
}
\description{
Describe the function
}
