% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AfsariCorVarcpp2}
\alias{AfsariCorVarcpp2}
\title{Describe the function}
\usage{
AfsariCorVarcpp2(mydist)
}
\arguments{
\item{describe}{the params}
}
\value{
describe the return
}
\description{
Describe the function
}
