% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REVA.R
\docType{package}
\name{REVA}
\alias{REVA}
\title{REVA: testing whther a sclalar value differs less for coler points}
\description{
We hava a scalra value for a set of points and we have a distance between the points.
Null ypothesis: the scalar does not depend ont the location, so the values differ between distant points as strons as between close points
Main function:
}
\section{REVA functions}{

REVA.p.value calculates the p-value corresponding to the null
REVA.user.distance calculate distance as Kendall tau and calculate the p-value	
REVA.user.distance.Geneset does something
}

