% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REVAFunction2Use.R
\name{REVA.test}
\alias{REVA.test}
\title{REVA.test}
\usage{
REVA.test(x, mydist, sampleNum2EstVar = 100, permsN = 10)
}
\description{
The main function of the REVA package. Calculates REVA statistics for a set of scalar values for each point (x) and the distances between points (mydist)
}
